pyproject.toml
[tool.poetry]
name = "rekenkamer-j-en-v-orgs"
version = "0.1.0"
description = ""
authors = ["Bing Steup <bingsteup@gmail.com>"]
readme = "README.md"
packages = [{include = "rekenkamer_j_en_v_orgs"}]

[tool.poetry.dependencies]
python = "^3.11"
elastic-enterprise-search = "^8.4.0"
openpyxl = "^3.1.2"
pandas = "^1.5.3"
minio = "^7.1.13"
python-dotenv = "^1.0.0"
ecs-logging = "^2.0.0"
pymongo = "^4.4.1"


[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-dotenv = "^0.5.2"
time-machine = "^2.9.0"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

